CHG: TrackDataCache.ClassifyLabels() allows not to detect for empty images
C++ version of the measures is strict, it always stops measuring whenever
either GT or RES empty image is found. The current behavior of the plugins
is the same. However, one has the option to disable this -- only in the
AOGM:Consistency checker plugin; the TRA always checks (cannot be disabled).